requirements.txt
python-telegram-bot
google-generativeai
feedparser
python-dotenv
flask
gspread
oauth2client
gunicorn
